Rename all files to match their SF; move Unlinking to own file
