+++ test/test_LinkKnowledgeElements.js
@@ -6,7 +6,6 @@ const {
   setUpJira,
   createJiraIssue,
   localCredentialsObject,
-  base64LocalCredentials,
   getKnowledgeElements,
   createDecisionKnowledgeElement,
 } = require('./helpers.js');
@@ -220,185 +219,6 @@ describe('TCS: Test link knowledge elements', () => {
       );
     } catch (err) {
       chai.expect(err.message).to.eql('Request failed with status code 400');
-    }
-  });
-});
-
-// TODO: move this to a separate file
-/**
- * CONDEC-172: Unlink knowledge elements
- */
-describe('TCS: Test unlink knowledge elements', () => {
-  before(async () => {
-    await setUpJira(true); // turn issue strategy on
-  });
-
-  /**
-   * TCS: Test unlink knowledge elements should delete the Jira issue link when both source and target elements of unlinking are Jira issues (R1)
-   *
-   * Precondition: Two decision knowledge elements exist as Jira issues and are
-   * linked to each other
-   *
-   * Step 1: Request the link between the issues to be deleted via the ConDec
-   * deleteLink interface
-   *
-   * Step 2: Verify that the Jira issue link between the elements has been removed
-   *
-   * Postcondition: The Jira issue link between the elements has been removed
-   */
-  it('should delete the Jira issue link when both source and target elements of unlinking are Jira issues (R1)', async () => {
-    const issue1 = await createJiraIssue('Issue', 'Issue 1');
-    const issue2 = await createJiraIssue('Alternative', 'Issue 2');
-
-    // Create a link between issue 1 and issue 2
-    await axios.post(
-      `${JSONConfig.fullUrl}/rest/condec/latest/knowledge/createLink.json` +
-        `?projectKey=${JSONConfig.projectKey}` +
-        '&documentationLocationOfParent=i' +
-        '&documentationLocationOfChild=i' +
-        `&idOfParent=${issue1.id}` +
-        `&idOfChild=${issue2.id}` +
-        '&linkTypeName=relates',
-      undefined,
-      localCredentialsObject
-    );
-
-    // Then delete the link
-    const payload = {
-      // THESE ELEMENTS MUST BE PASSED IN THIS ORDER!!
-      idOfSourceElement: issue1.id,
-      idOfDestinationElement: issue2.id,
-      documentationLocationOfSourceElement: 'i',
-      documentationLocationOfDestinationElement: 'i',
-    };
-    const deleteLinkRequest = {
-      method: 'delete',
-      url:
-        `${JSONConfig.fullUrl}/rest/condec/latest/knowledge/deleteLink.json` +
-        `?projectKey=${JSONConfig.projectKey}`,
-      headers: {
-        Authorization: `Basic ${base64LocalCredentials}`,
-        'Content-Type': 'application/json',
-      },
-      data: payload,
-    };
-    try {
-      // Due to some strangeness with axios, we have to construct the request as above before
-      // sending it
-      const deleted = await axios(deleteLinkRequest);
-      chai.expect(deleted.status).to.eql(200);
-    } catch (err) {
-      console.error(err);
-    }
-    const searchResult = await jira.findIssue(issue1.key);
-    // Make sure it's really gone
-    chai
-      .expect(searchResult.fields.issuelinks)
-      .to.be.an('Array')
-      .with.length(0);
-  });
-
-  /**
-   * TCS: Test unlink knowledge elements should set the status of an issue to "unresolved" when it is unlinked from a "decided" decision and has no other "decided" decisions linked (R3)
-   *
-   * Precondition: A decision knowledge issue exists and is linked to a decision
-   * in status "decided"
-   *
-   * Step 1: Delete the link between the knowledge issue and the decision
-   *
-   * Step 2: Verify that the issue now has status "unresolved"
-   *
-   * Postcondition: The issue's status is "unresolved"
-   */
-  // This test is currently failing, it seems as if the issue gets deleted and
-  // not just unlinked
-  it('should set the status of an issue to "unresolved" when it is unlinked from a "decided" decision and has no other "decided" decisions linked (R3)', async () => {
-    const issue = await createDecisionKnowledgeElement(
-      'Which qualifications should be considered in hiring a new developer?',
-      'Issue',
-      'i'
-    );
-
-    const comment = await createDecisionKnowledgeElement(
-      'Consider the amount of experience the candidate has!',
-      'Decision',
-      's',
-      issue.id,
-      'i'
-    );
-    const payload = {
-      idOfSourceElement: comment.id,
-      idOfDestinationElement: issue.id,
-      documentationLocationOfSourceElement: 's',
-      documentationLocationOfDestinationElement: 'i',
-    };
-    const deleteLinkRequest = {
-      method: 'delete',
-      url:
-        `${JSONConfig.fullUrl}/rest/condec/latest/knowledge/deleteLink.json` +
-        `?projectKey=${JSONConfig.projectKey}`,
-      headers: {
-        Authorization: `Basic ${base64LocalCredentials}`,
-        'Content-Type': 'application/json',
-      },
-      data: payload,
-    };
-    const response = await axios.request(deleteLinkRequest);
-    chai.expect(response.status).to.eql(200);
-    const knowledgeElements = await getKnowledgeElements();
-    chai
-      .expect(knowledgeElements)
-      .to.be.an('Array')
-      .that.contains.something.like({ id: issue.id, status: 'unresolved' });
-  });
-
-  /**
-   * TCS: Test unlink knowledge elements should not allow unlinking of elements with nonexistent ids (E1)
-   *
-   * Precondition: Two decision knowledge elements exist and are not linked to
-   * each other
-   *
-   * Step 1: Attempt to trigger deletion of the link between the two elements
-   *
-   * Step 2: Verify that the deletion failed
-   *
-   * Postcondition: Nothing changed
-   */
-  it("should not allow unlinking of a link that doesn't exist (E1)", async () => {
-    const issue = await createDecisionKnowledgeElement(
-      'How should files be organized?',
-      'Issue',
-      'i'
-    );
-    // Create a decision that is not linked to the issue
-    const decision = await createDecisionKnowledgeElement(
-      'Organize files alphabetically!',
-      'Decision',
-      'i'
-    );
-
-    const payload = {
-      // THESE ELEMENTS MUST BE PASSED IN THIS ORDER!!
-      idOfSourceElement: issue.id,
-      idOfDestinationElement: decision.id,
-      documentationLocationOfSourceElement: 'i',
-      documentationLocationOfDestinationElement: 'i',
-    };
-    const deleteLinkRequest = {
-      method: 'delete',
-      url:
-        `${JSONConfig.fullUrl}/rest/condec/latest/knowledge/deleteLink.json` +
-        `?projectKey=${JSONConfig.projectKey}`,
-      headers: {
-        Authorization: `Basic ${base64LocalCredentials}`,
-        'Content-Type': 'application/json',
-      },
-      data: payload,
-    };
-    try {
-      await axios(deleteLinkRequest);
-    } catch (err) {
-      chai.assert(err);
     }
   });
 });
